requirements.txt
addict==2.4.0
agents-sdk==0.0.0
annotated-types==0.7.0
anthropic==0.72.0
antlr4-python3-runtime==4.13.2
anyio==4.11.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
arxiv==2.2.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
autograd==1.8.0
autoray==0.8.0
azure-ai-documentintelligence==1.0.2
azure-core==1.36.0
azure-identity==1.25.1
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cirq==1.6.1
cirq-aqt==1.6.1
cirq-core==1.6.1
cirq-google==1.6.1
cirq-ionq==1.6.1
cirq-pasqal==1.6.1
cirq-web==1.6.1
click==8.3.0
cobble==0.1.4
coloredlogs==15.0.1
colour==0.1.5
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Cython==3.1.6
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
diastatic-malt==2.15.2
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.4.1
duet==0.2.9
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
feedparser==6.0.12
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
gast==0.6.0
glcontext==3.0.0
google-api-core==2.28.1
google-auth==2.42.0
google-genai==1.46.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.70.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isoduration==20.11.0
isosurfaces==0.1.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
langchain==1.0.2
langchain-core==1.0.1
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
lark==1.3.1
lxml==6.0.2
magika==0.6.2
mammoth==1.10.0
manimgl==1.7.2
ManimPango==0.6.1
mapbox_earcut==1.0.3
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.3
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
narwhals==2.10.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook_shim==0.2.4
numpy==2.3.4
olefile==0.47
onnxruntime==1.23.2
openai==2.6.1
openpyxl==3.1.5
openqasm3==1.0.1
opentelemetry-api==1.38.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pdfminer.six==20250506
pennylane==0.43.0
pennylane_lightning==0.43.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
polars==1.34.0
polars-runtime-32==1.34.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyautogen==0.10.0
pycparser==2.23
pydantic==2.11.1
pydantic_core==2.33.0
pydub==0.25.1
pyglet==2.1.9
pyglm==2.8.2
Pygments==2.19.2
PyJWT==2.10.1
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
PyOpenGL==3.1.10
pyparsing==3.2.5
pyperclip==1.11.0
pyqasm==1.0.0
pyspnego==0.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qbraid==0.10.0
qbraid-core==0.1.45
qiskit==2.2.2
qiskit-ibm-runtime==0.43.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
rustworkx==0.17.1
scipy==1.16.3
scipy-openblas32==0.3.30.0.4
screeninfo==0.8.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SpeechRecognition==3.14.3
stack-data==0.6.3
stevedore==5.5.0
svgelements==1.9.6
sympy==1.14.0
tavily-python==0.7.12
tenacity==9.1.2
termcolor==3.2.0
terminado==0.18.1
tiktoken==0.12.0
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typedunits==0.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uv==0.9.5
validators==0.35.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
youtube-transcript-api==1.0.3
zipp==3.23.0
zstandard==0.25.0
